Remove c-style cast

+++ src/mandelbrot.cpp
@@ -24,7 +24,7 @@ static auto Color(const int iterations)
     const auto val = (g_max_iterations == iterations) ? 0.0f : 1.0f;
 
     const auto h = hue / 60;
-    const auto f = (float)hue / 60 - h;
+    const auto f = hue / 60.0f - h;
     const auto p = val * (1.0f - sat);
     const auto q = val * (1.0f - sat * f);
     const auto t = val * (1.0f - sat * (1.0f - f));
